FOTA notification improvement when device reboots
When a user agreement notification is sent to the application
(LE_AVC_DOWNLOAD_PENDING, LE_AVC_INSTALL_PENDING), if the application
does not accept the agreement and if the device reboots, the
notification is sent to the application at the next connection to
the server.
Moreover, for a FOTA, when the device reboots in order to install
the downloaded package, the LE_AVC_INSTALL_COMPLETE/FAILED notification
is sent to the application at the next connection to the server.
All these notifications need to be sent when the application
registers to the airVantageConnector event handler.

Resolves: LE-6933, LE-6985, LE-6953
Change-Id: Ia07c9d5ebbb9f25f079d4310f26546bc218a3316