Launch FW update FOTA process only after user agreement
The FW update process has a 15 minutes timeout and is now launched
only when the user agreed to download the FOTA update package.
This avoids the timeout expiration and therefore the FOTA process
failure in case the user does not agree to the download and just waits
after receiving the DOWNLOAD_PENDING notification.

Resolves: LE-7157
Change-Id: I44be000716d04ee05b9da397f25cb49010a0e0e6