Connection notification to airVantage server
After a FOTA install (platform reboot), the platform does not
automatically connect to the airVantage server to notify the install
result.
A new LE_AVC_CONNECTION_REQUIRED notification is added in order to
notify the customer application which controls airVantage services that
a connection is required to the airVantage server.

Resolves: LE-6423
Change-Id: I15c5a9828c67d655785aab8aa82bc3cd1ad3065c