Use 3rdParty-provided tinycbor
At the moment, tinycbor is expected to be provided by the toolchain.
This makes it necessary to provide a customized toolchain to be able
to build/use AVC.

Instead this uses a tinycbor built through an externalBuild component
directly from sources.

The tinyCBOR build system only provides a static library (.a) and no
dynamic library (.so). The existing .so in the toolchain was generated
through a patch in our Yocto distribution.

Resolves: LE-6565
Change-Id: I5dce1e74f463d73e5ee64b212f8253381c1a33c6