Fix and revise behaviour for LE_AVC_NO_UPDATE notification
LE_AVC_NO_UPDATE should not be returned when download pending and
install pending are received. Only time it should be received is
when there is an expected  interaction between the device and server side.
The two cases being:
- When device start a session with server to check if there's a job
- When download is complete and device waits for server to send install cmd

Also fix issue where LE_AVC_NO_UPDATE is not returned when download complete/
failed and no install pending is received.

Resolves: LE-7680
Change-Id: I364e018e7ef6c5c0465cc593da1d0efc5e20cd2f