Fix polling timer disabled value configuration
Lwm2mcore defines the polling timer disabled value to 630720000s which
is greater than LE_AVC_POLLING_TIMER_MAX_VAL 31536000s. This is seen
as an out of range value. Simply check for the lwm2mcore default value
and set avc's polling timer configuration value to disabled in that
case.

Resolves: LE-7937
Change-Id: Iab4ee1184e67e21273cb830c2a49ef4b78209373