Fix for SMS not being deleted by AVC
The problem is that AVC is deleting SMS only if it is
identified to be a Wake-Up SMS.
The fix:
Since the Delete() call is just a releasing/dereferencing,
AVC should always invoke it after the message is processed. The
message will be deleted after it's released by all clients.
Also contains improvement of the logic checking the message
prefix.

Resolves: LE-11505
Change-Id: I720f0cfaaa19c5e79002130502c4f44cc340f1e3