WIFI intermittently displayed as Technology on AVMS
The problem is that le_data_GetTechnology() may return WIFI when
DCS is in the process of attempting WiFi connection (for example
due to unstable cellular connection).
The fix is to make lwm2mcore_GetNetworkBearer() (and other APIs
from the same group) return INVALID_STATE if connection is down.
In this case, if the system is in "attempting connection" state,
the synchronize job will still return without error, but the
technology field will not be received from the device, and
will not be updated in the server GUI.

Resolves: LE-11782
Change-Id: I88a3a93750cc9708ff73232bb35512e1bc0dd8bd