Fix unused values error on Travis CI + remove osPortSecurity.c
It seems that it complains in the travis environment:
/home/travis/build/legatoproject/workspace/legato/apps/platformServices/airVantageConnector/avcClient/osPortSecurityOpenSSL.c: In function ‘lwm2mcore_Base64Encode’:
/home/travis/build/legatoproject/workspace/legato/apps/platformServices/airVantageConnector/avcClient/osPortSecurityOpenSSL.c:118:5: error: value computed is not used [-Werror=unused-value]
     BIO_flush(bio);
     ^
/home/travis/build/legatoproject/workspace/legato/apps/platformServices/airVantageConnector/avcClient/osPortSecurityOpenSSL.c:120:5: error: value computed is not used [-Werror=unused-value]
     BIO_set_close(bio, BIO_NOCLOSE);
     ^
cc1: all warnings being treated as errors
ninja: build stopped: subcommand failed.
make: *** [system] Error 1

The error is very similar to I82c269b45a348d68a0c5b46e91890444eecfbd24 ,
so fixing it the exact same way.
The original file it was applied to (osPortSecurity.c) had been renamed
on the branch but without the fix.
When the change got introduced on the master branch, the original file
was kept while it is now dead code, so removing it.

Resolves: LE-12739
Related: LE-11220
Change-Id: I390ba4aa8de13089affec32cd23f7e19ac2ea5e5