Filter some notifications
When REGISTER UPDATE tries to be sent with DTLS resume while the
device is connected to the DM server, if DTLS retransmission are
made, some notification could be returned to the registered
applications.
Such notifications are not necessary and should be filtered.

Resolves: LE-14120 LE-14126
Change-Id: Iaceaca0a57f37fd18fdff349dea4ae5ed1607e1f
(cherry-pick from commit debea1c5f7d8a79374e25fc281bd85dfa196a675)