Fix connectivity on package download end
When package download ends, avcClient_Update function is called to
check the route between the device and the server.
This function is available if the device is connected to the server
at that moment.
Initiate a connection to the server in this case if the device is not
connected to the server.

Resolves: LE-13840
Change-Id: I7365b26fe1b2c991ea268eb37e6243a1340df5ef
Relates: LE-13216