Add le_avc_SetNatTimeout API
Add le_avc_SetNatTimeout API to set the NAT timeout in the AV client. Previously, this NAT
timeout value was set to 40 seconds.
When data need to be sent by the client, a check is made between this NAT timeout value and the
time when last data were received from the server or sent to the server.
If one of these times is greater than the NAT timeout, a DTLS resume is initiated.
Default value if this function is not called: 40 seconds.
Value 0 will deactivate any DTLS resume.
Value is stored in volatile memory.

Resolves: LE-15648
Change-Id: Ib911b9552f4c11ad29ec901f1c6e35291e71e317