Fix AVC connectivity issue when data session is disconnected
When the data session is disconnected due to poor network or explicitly turned off. Upon resuming, the AVC connection would sometimes fail. This would affect SOTA/FOTA suspend and resume use cases where the connection will resume immediately and fail to connect. This issue was introduced by the enhancement of performing dtls resume instead of a full handshake when an AVC session is disconnected. However as a result, when the data session is disconnected, we need to perform a dtls resume instead of relying on the dtls nat timeout. Reducing the dtls nat timeout to 5s in this specific scenario should ensure that a dtls resume would occur.

Resolves: LE-16590
Change-Id: I55cb1973a79e9f64fa30342767e158ffcc95ebe8